refactor likes into ratings
* refactor to use new Rating model to store likes (will be changed to full-fledged ratings in the future)

* add necessary data access to DbLikesStorage to accommodate for recommendation system

* placeholders for recommendation service components

* add data access for diff and freq matrices

* create recommendation system business logic

* implement recommendations endpoint, auto-updates for diff and freq, and RecommendationControllerTest

* update Slope One logic to correctly average diff values

* refactor liking endpoints into rating endpoints

* films are now fetched along with their avg ratings; fetch popular films based on avg rating

* update FilmService and FilmStorage to work with new data types

* update film data extractor to fetch director data

* add back the option to add "likes"

* rename parameter for getting top films by director

* make handling legacy likes more concise

* switch ratings from int to double for some reason

* default to fetching all popular films when count not specified

* when recommending, filter out films with avg rating not higher than 5.0

* update recommendation endpoint tests

* translate log & error messages

---------

+++ src/main/java/ru/yandex/practicum/filmorate/model/Rating.java
@@ -10,6 +10,6 @@ public class Rating {
 
     private long userId;
     private long filmId;
-    private int ratingValue;
+    private double ratingValue;
     private Timestamp ratedAt;
 }
